Create get_model.py
Loads a model from an HDF5 file (architecture, weights and compilation or learning configuration). The model instantiated in this file based on the schematic for the convolutional neural network saved in the HDF5 file is used to predict whether an image of a cat or a dog is processed in the batch retrieved using the data retrieving and formatting object employed (cats_and_dogs(10, 0): 10 images for first parameter).